LayoutTests/platform/win/http/tests/xmlhttprequest/send-array-buffer-expected.txt

CONSOLE MESSAGE: line 18: ReferenceError: Can't find variable: Uint8Array
Tests sending an array buffer.
